scraping/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
urllib3==2.1.0
Pillow==10.1.0
tqdm==4.66.1
